clean up, consolidate ng metadata gen
